pyproject.toml
[tool.poetry]
name = "google-map-analyzer"
version = "0.1.0"
description = ""
authors = ["TomoyaFujita2016 <fujita.t.2016c@gmail.com>"]
readme = "README.md"
packages = [{include = "google_map_analyzer"}]

[tool.poetry.dependencies]
python = ">=3.9.2,<3.9.7 || >3.9.7,<4.0.0"
streamlit = "^1.23.1"
requests = "^2.31.0"
tqdm = "^4.65.0"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
